requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
waitress==3.0.0
